+++ crates/moon/tests/test_cases/hello/main/main.mbt
