+++ crates/moon/tests/test_cases/hello/main/moon.pkg.json
