+++ crates/moon/tests/test_cases/moon_commands/lib/list/moon.pkg.json
