+++ crates/moon/tests/test_cases/moon_commands/lib/queue/moon.pkg.json
