+++ crates/moon/tests/test_cases/moon_commands/lib/stack/moon.pkg.json
