+++ crates/moon/tests/test_cases/moon_commands/lib/vec/lib.mbt
