+++ crates/moon/tests/test_cases/moon_commands/lib/vec/moon.pkg.json
