+++ crates/moon/tests/test_cases/moon_commands/main1/main.mbt
