+++ crates/moon/tests/test_cases/moon_commands/main1/moon.pkg.json
