+++ crates/moon/tests/test_cases/moon_commands/main2/moon.pkg.json
